+++ src/components/atoms/ImageUpload/index.tsx
@@ -1,6 +1,7 @@
 import React, { useState } from "react";
 import { Button, Box, Avatar } from "@chakra-ui/react";
-
+import { colors } from "@src/theme/colors";
+import styled from "styled-components";
 interface IProps {
   setValue?: any;
   required?: boolean;
@@ -9,6 +10,24 @@ interface IProps {
   name?: string;
 }
 
+const ImageStyled = styled.div`
+  display: flex;
+  border-radius: 7px;
+  border: solid 1px ${colors.light_gray_border};
+  color: ${colors.text};
+  label {
+    border-right: solid 1px ${colors.light_gray_border};
+    padding: 10px;
+    background-color: ${colors.light_white};
+    cursor: pointer;
+  }
+  p {
+    padding: 10px;
+  }
+  #fileInput {
+    display: none;
+  }
+`;
 const ImageUpload: React.FC<IProps> = ({
   setValue,
   required,
@@ -17,15 +36,15 @@ const ImageUpload: React.FC<IProps> = ({
   name = "image",
 }) => {
   const [selectedImage, setSelectedImage] = useState<any>(null);
-
+  const [imageName, setImageName] = useState("");
   const handleImageChange = (event: any) => {
     const selectedImage = event?.target?.files[0];
     if (selectedImage) {
       setSelectedImage(URL.createObjectURL(selectedImage));
+      setImageName(selectedImage?.name);
       setValue(name, event.target.files[0]);
     }
   };
-
   return (
     <Box textAlign="center">
       {isUser && (
@@ -66,15 +85,19 @@ const ImageUpload: React.FC<IProps> = ({
           </Button>
         </>
       )}
-
-      <input
-        id="fileInput"
-        type="file"
-        onChange={handleImageChange}
-        accept="image/*"
-        style={{ display: isUser ? "none" : "block" }}
-        required={required}
-      />
+      <ImageStyled>
+        <label htmlFor="fileInput">Choose a file</label>
+        <input
+          id="fileInput"
+          type="file"
+          onChange={handleImageChange}
+          accept="image/*"
+          required={required}
+        />
+        <p id="selectedFileName">
+          {imageName ? imageName : "No file selected"}
+        </p>
+      </ImageStyled>
     </Box>
   );
 };